requirements.txt
confluent_kafka==1.8.2
google-cloud-pubsub==2.12.0